* Cleanup, use longs rather than ints where appropriate * Re-enabled SoundTouch code, if Mixxx crashes on you and you figure out the bug in lines 174-192 (enginebufferscalest.cpp), I will give you $5 * Bonus: I will give you an extra $5 if I'm wrong and the bug isn't where I said it is * Added ASSERT to EngineBufferScaleST to help track down problems

+++ mixxx/src/enginebufferscalelinear.h
@@ -28,7 +28,7 @@ class EngineBufferScaleLinear : public EngineBufferScale  {
 public:
     EngineBufferScaleLinear(ReaderExtractWave *wave);
     ~EngineBufferScaleLinear();
-    CSAMPLE *scale(double playpos, int buf_size, float *pBase, int iBaseLength);
+    CSAMPLE *scale(double playpos, unsigned long buf_size, float *pBase, unsigned long iBaseLength);
     void setBaseRate(double dBaseRate);
     double setTempo(double dTempo);
     void clear();
